Add files via upload
This project is basically on detecting the lung cancer disease by using ML's algorithm like KNN & DECISION TREE and also with the help of Scikit-learn libraries like Sklearn and pandas to predict and classify the dataset of the lung cancer patients. Slicing the dataset  and feature scaling options are used to train the dataset. After that we use confusion matrix , f1 score and accuracy score to predict the accuracy of the result.